Refact : 리팩토링
