+++ src/core/api.ts
@@ -1,0 +1,35 @@
+import { NewsFeed,NewsDetail } from "../types";
+
+
+export class Api {
+  ajax: XMLHttpRequest;
+  url: string;
+  constructor(url: string) {
+    this.ajax = new XMLHttpRequest();
+    this.url = url;
+  }
+  getRequest<AjaxResponse>(): AjaxResponse {
+    this.ajax.open('GET', this.url, false);
+    this.ajax.send();
+    return JSON.parse(this.ajax.response)
+  }
+}
+
+export class NewsFeedApi extends Api {
+  constructor(url : string) {
+    super(url)
+  }
+  getData(): NewsFeed[] {
+    return this.getRequest<NewsFeed[]>();
+  }
+}
+
+
+export class NewsDetailApi extends Api{
+  constructor(url: string) {
+    super(url);
+  }
+  getData(): NewsDetail {
+    return this.getRequest<NewsDetail>();
+  }
+}
