+++ src/core/store.ts
@@ -1,0 +1,7 @@
+import { Store } from "../types";
+
+export const store: Store = {
+    currentPage: 1,
+    feeds : new Map()
+  };
+
